Consolidate controller bootstrap login in `run` command
I noticed that externally provided certificate is actually getting copied upon `orchard controller init` which led me to thinking maybe it's not the right concept to have a separate `init` command.

This change proposes to move the bootstrap to `run` command:
1. If external certificate provided then just use it without coping.
2. Otherwise generate a self-signed one.
3. If a "superuser" is provided, then make sure we update it in the database.

This way users won't need to run `init` first. One will be able to generate a token for the superuser, configure the run arguments using it. Then reuse the superuser credentials to generate registration tokens for workers and start working with the API.